refactor(dashboard): share readTextTail via fs-utils for maintainability
Sonar maintainability B on new code; centralize tail-byte read helper
used by dashboard, feedback-loop, and aggregate JSONL paths.
